% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_model.R
\name{aggregate_tiers}
\alias{aggregate_tiers}
\title{Implement the tiered_factor_logic on all factors}
\usage{
aggregate_tiers(flight_schedule, factors)
}
\arguments{
\item{flight_schedule}{dataframe}

\item{factors}{character}
}
\value{
dataframe
}
\description{
Implement the tiered_factor_logic on all factors
}
