% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packages.R
\name{assign_mod_inputs}
\alias{assign_mod_inputs}
\title{List of Model Inputs}
\usage{
assign_mod_inputs(dat_pack, minute_bin = 5)
}
\arguments{
\item{dat_pack}{DB Object}

\item{minute_bin}{integer}
}
\description{
List of Model Inputs
}
