% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_model.R
\name{bag_model}
\alias{bag_model}
\title{Main function for Bag Model}
\usage{
bag_model(mod_inputs, minute_bin)
}
\arguments{
\item{mod_inputs}{list: all data packages}

\item{minute_bin}{integer: bin the departure times}
}
\value{
dataframe
}
\description{
Take the base of a flight schedule, join appropriate arrival curves.
Aggregate and join all the model factors. For OA's, the df_oa_local
dataframe is used for orig_factor percentages. Filter out what carriers
are needed based on shared_checkpoints dataframe.
Finally, create expected demand columns.
}
