% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caro_model.R
\name{carousel_model}
\alias{carousel_model}
\title{Carousel model}
\usage{
carousel_model(df_bag_model, minutes_prior_bag_pickup = 120)
}
\arguments{
\item{df_bag_model}{dataframe}

\item{minutes_prior_bag_pickup}{integer}
}
\description{
Carousel model
}
