% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint_model.R
\name{checkpoint_model}
\alias{checkpoint_model}
\title{Checkpoint Model}
\usage{
checkpoint_model(df_pax_model, mod_inputs, minute_bin = 5)
}
\arguments{
\item{df_pax_model}{dataframe}

\item{mod_inputs}{list}

\item{minute_bin}{integer}
}
\description{
Checkpoint Model
}
