% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packages.R
\name{clean_arrival_curve}
\alias{clean_arrival_curve}
\title{Group arrivals and get percentage of PAX arriving in given time
bin prior to departure. Arranges vector into a wide format to join
onto each flight.}
\usage{
clean_arrival_curve(data, minute_bin = 5)
}
\arguments{
\item{data}{dataframe}

\item{minute_bin}{integer: bin the departure time}
}
\description{
Group arrivals and get percentage of PAX arriving in given time
bin prior to departure. Arranges vector into a wide format to join
onto each flight.
}
