% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packages.R
\name{clean_load_factors}
\alias{clean_load_factors}
\title{Clean load factors from DB, filter out for orig-dest-dep hour
pairings that have less than 2 flights}
\usage{
clean_load_factors(data, minute_bin = 10)
}
\arguments{
\item{data}{tibble: of load factors from DB}

\item{minute_bin}{integer: bin the dep_int}
}
\description{
Clean load factors from DB, filter out for orig-dest-dep hour
pairings that have less than 2 flights
}
