% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packages.R
\name{dbConnectSQLite}
\alias{dbConnectSQLite}
\title{Connect to data packages in SQLite DB}
\usage{
dbConnectSQLite(SQLiteDB)
}
\arguments{
\item{SQLiteDB}{A string}
}
\value{
DB Connection
}
\description{
Connect to data packages in SQLite DB
}
