% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_oa_local}
\alias{df_oa_local}
\title{Other Airline Local Markets}
\format{
A dataframe with 1507 rows and 8 variables:
\describe{
\item{dominant_marketing_airline}{Carrier}
\item{origin}{origin}
\item{passengers_per_day}{Total Passengers}
\item{connection_points}{Route pairing}
\item{right_dominant_marketing_airline}{Other carrier}
\item{right_passengers_per_day}{Joined route passengers}
\item{total}{total}
\item{local}{local}
}
}
\usage{
df_oa_local
}
\description{
A dataset from 2019 providing the local factor of passengers
}
\keyword{datasets}
