% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pgds}
\alias{df_pgds}
\title{PGDS Arrival Curve}
\format{
A dataframe with 5 rows and 18 variables:
\describe{
\item{distance}{international or domestic}
\item{timeframe}{timeframe before or after 9 am}
\item{perc_bin10}{increment of pax arriving at 1 min prior}
\item{perc_bin20}{increment of pax arriving at 1 min prior}
\item{perc_bin30}{increment of pax arriving at 1 min prior}
\item{perc_bin40}{increment of pax arriving at 1 min prior}
\item{perc_bin50}{increment of pax arriving at 1 min prior}
\item{perc_bin60}{increment of pax arriving at 1 min prior}
\item{perc_bin70}{increment of pax arriving at 1 min prior}
\item{perc_bin80}{increment of pax arriving at 1 min prior}
\item{perc_bin90}{increment of pax arriving at 1 min prior}
\item{perc_bin100}{increment of pax arriving at 1 min prior}
\item{perc_bin110}{increment of pax arriving at 1 min prior}
\item{perc_bin120}{increment of pax arriving at 1 min prior}
\item{perc_bin130}{increment of pax arriving at 1 min prior}
\item{perc_bin140}{increment of pax arriving at 1 min prior}
\item{perc_bin150}{increment of pax arriving at 1 min prior}
\item{perc_bin160}{increment of pax arriving at 1 min prior}
}
}
\usage{
df_pgds
}
\description{
A dataset that provides a standard arrival curve to use for
other airlines
}
\keyword{datasets}
