% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_transaction}
\alias{df_transaction}
\title{Average Transaction Times for Ticketing}
\format{
A dataset with 68 rows and 3 variables
\describe{
\item{station}{origin}
\item{curbside_transaction_time}{skycap service rate}
\item{ticketing_transaction_time}{full service service rate}
}
}
\usage{
df_transaction
}
\description{
Average Transaction Times for Ticketing
}
\keyword{datasets}
