% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pax_model.R
\name{distribute_demand}
\alias{distribute_demand}
\title{Distribute Pax and Bag Demand}
\usage{
distribute_demand(df_bag_model, vector_name, col_name)
}
\arguments{
\item{df_bag_model}{dataframe}

\item{vector_name}{character}

\item{col_name}{character}
}
\description{
Take the vector of expected demands, and multiply them across the
matrix of arrival curve percentages. This places the proportion of expected
demand in the given minutes prior increments. The data is pivoted to
subtract the minutes prior from departure datetime, and summarized by
the appropriate times. This returns a single time increment for each origin.
}
