% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_model.R
\name{gate_model}
\alias{gate_model}
\title{Gate Model}
\usage{
gate_model(mod_inputs, minute_bin = 5)
}
\arguments{
\item{mod_inputs}{list}

\item{minute_bin}{integer}
}
\description{
Models the planes occupying gates during the day. After classifying plane types, add
buffer time to departure time. From there work, the way back of occupying a gate from dep time
\itemize{
\item buffer, to the arrival time. Arrival times that are NULL are considered originators. Originator
logic is calculated by taking the cumsum of of binary orignator values. This will give which number
originator the plane is.
}
}
