% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packages.R
\name{generate_station_assumptions}
\alias{generate_station_assumptions}
\title{Pull in assumptions from quickbase}
\usage{
generate_station_assumptions()
}
\value{
dataframe
}
\description{
Pull in assumptions from quickbase
}
