% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_model.R
\name{join_flight_sched_arr_curve}
\alias{join_flight_sched_arr_curve}
\title{Join arrival curve to flight schedule}
\usage{
join_flight_sched_arr_curve(mod_inputs, minute_bin)
}
\arguments{
\item{mod_inputs}{list: all data packages}

\item{minute_bin}{integer: bin the departure times}
}
\value{
dataframe
}
\description{
Join arrival curve to flight schedule
}
