% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_models.R
\name{main_model}
\alias{main_model}
\title{Main Function}
\usage{
main_model(mod_inputs, minute_bin = 5)
}
\arguments{
\item{mod_inputs}{list}

\item{minute_bin}{integer}
}
\description{
Runs all the models and joins them together for a typical model run
}
