% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pax_model.R
\name{pax_model}
\alias{pax_model}
\title{Pax Modle}
\usage{
pax_model(df_bag_model)
}
\arguments{
\item{df_bag_model}{dataframe}
}
\description{
Applies the distribute_demand to each vector and creates a joined output.
}
