% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_station_model.R
\name{service_station_model}
\alias{service_station_model}
\title{Service Station Model}
\usage{
service_station_model(df_pax_model, mod_inputs, minute_bin = 5)
}
\arguments{
\item{df_pax_model}{dataframe}

\item{mod_inputs}{list: object of data packages}

\item{minute_bin}{integer}
}
\description{
Model lobby areas of Full Service, Skycap, and STK's as well as bag drop.
Major room for imporvements here, but for now it uses basic queue theoy
deterministic logic. Service Rate is how many PAX can be processed in the
binned time increment. This function processes the pax, gives the queues and wait times
at each time increment.
}
