% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_model.R
\name{tiered_factor_logic}
\alias{tiered_factor_logic}
\title{Tiered Factors}
\usage{
tiered_factor_logic(df)
}
\arguments{
\item{df}{dataframe: flight schedule}
}
\value{
dataframe
}
\description{
Create a grouping by orig, dest and dep hour and use
what is available.
}
