% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpoint_model.R
\name{tpoint_model}
\alias{tpoint_model}
\title{Tpoint Model}
\usage{
tpoint_model(df_bag_model, cart_staging_time = 120, minute_bin = 5)
}
\arguments{
\item{df_bag_model}{dataframe}

\item{cart_staging_time}{integer}

\item{minute_bin}{integer}
}
\description{
Similar to how the arrival curve is implemented, a tpoint curve is created from
30 min prior to departure up to the cart staging time minutes prior to departure.
A cart represents 45 bags, we cap it at 3 carts for a flight. Carts are then allocated
based on the number required for the flight and how long the cart staging time is.
At 1 hour prior to departure, a cart is removed from the allocation.
}
\details{
If a flight is expecting 100 bags, that would represent 3 carts due to a ceiling
function. At a stage time of 120 min, 3 carts would be allocated 2 hours before departure.
They would remain there until an hour prior, at which time 1 cart would be removed. The
remaining 2 carts would then be allocated at tpoint until 30 min prior to departure, at
which time both carts are removed.
}
