% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packages.R
\name{update_new_stations}
\alias{update_new_stations}
\title{Proxy logic for new stations to use paired station historical data}
\usage{
update_new_stations(data)
}
\arguments{
\item{data}{dataframe: col required 'orig'}
}
\description{
Proxy logic for new stations to use paired station historical data
}
